cambio de codigo

+++ backend/clinica/src/main/java/com/Clinica/Repository/OdontogramaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface OdontogramaRepository extends JpaRepository<Odontograma, Integer> {
-    List<Odontograma> findByPaciente_IdPaciente(Integer idPaciente);
+    List<Odontograma> findByPaciente_IdPaciente(long idPaciente);
 }
